Update specs
The HeadersSpec is still broken: we can no longer distinguish between recurring and non-recurring when headers are in raw form.
