+++ core/src/test/scala/org/http4s/headers/AcceptCharsetSpec.scala
@@ -1,17 +1,13 @@
-package org.http4s.parser
+package org.http4s.headers
 
-import org.http4s.headers.`Accept-Charset`
 import org.http4s._
-import scalaz.Validation
 
-class AcceptCharsetSpec extends Http4sSpec with HeaderParserHelper[`Accept-Charset`] {
-
-  def hparse(value: String): ParseResult[`Accept-Charset`] = HttpParser.ACCEPT_CHARSET(value)
+class AcceptCharsetSpec extends HeaderParserSpec(`Accept-Charset`) {
 
   "Accept-Charset" should {
     "parse any list of CharsetRanges to itself" in {
       prop { h: `Accept-Charset` =>
-        hparse(h.value) must beRightDisjunction(h)
+        hparse(h.value) must_== Some(h)
       }
     }
