+++ core/src/test/scala/org/http4s/headers/AllowSpec.scala
@@ -1,0 +1,13 @@
+package org.http4s
+package headers
+
+
+class AllowSpec extends HeaderParserSpec(Allow) {
+
+  "Allow Header" should {
+    "parse Allow" in {
+      val header = Allow(Method.GET, Method.POST)
+      hparse(header.value) must_== Some(header)
+    }
+  }
+}
