+++ core/src/test/scala/org/http4s/headers/AuthorizationSpec.scala
@@ -1,25 +1,21 @@
-package org.http4s.parser
-
-import org.http4s.{Http4sSpec, OAuth2BearerToken}
-import org.http4s.headers.Authorization
+package org.http4s
+package headers
 
 
-class AuthorizationHeaderSpec extends Http4sSpec {
-
-  def hparse(value: String) = HttpParser.AUTHORIZATION(value)
+class AuthorizationSpec extends HeaderParserSpec(Authorization) {
 
   "Authorization header" should {
     "Parse a valid Oauth2 header" in {
       val token = (('a' to 'z') ++ ('A' to 'Z') ++ ('0' to '9') ++ "-._~+/".toSeq).mkString
       val h = Authorization(OAuth2BearerToken(token + "="))
-      hparse(h.value) must beRightDisjunction(h)
+      hparse(h.value) must_== Some(h)
     }
 
-    "Reject an ivalid Oauth2 header" in {
+    "Reject an invalid Oauth2 header" in {
       val invalidTokens = Seq("f!@", "=abc", "abc d")
       forall(invalidTokens) { token =>
         val h = Authorization(OAuth2BearerToken(token))
-        hparse(h.value) must beLeftDisjunction
+        hparse(h.value) must_== None
       }
     }
   }
