+++ core/src/test/scala/org/http4s/headers/CacheControlSpec.scala
@@ -1,10 +1,7 @@
 package org.http4s
-package parser
+package headers
 
-import org.http4s.headers.`Cache-Control`
-import org.specs2.mutable.Specification
 import org.specs2.time.NoTimeConversions
-import scalaz.Validation
 import org.http4s.CacheDirective._
 import org.http4s.CacheDirective.`s-maxage`
 import org.http4s.CacheDirective.`max-stale`
@@ -12,10 +9,8 @@ import org.http4s.CacheDirective.`min-fresh`
 import org.http4s.CacheDirective.`max-age`
 import org.http4s.CacheDirective.`private`
 import scala.concurrent.duration._
-import org.http4s.util.string._
 
-class CacheControlSpec extends Specification with HeaderParserHelper[`Cache-Control`] with NoTimeConversions {
-  def hparse(value: String): ParseResult[`Cache-Control`] = HttpParser.CACHE_CONTROL(value)
+class CacheControlSpec extends HeaderParserSpec(`Cache-Control`) with NoTimeConversions {
 
   // Default values
   val valueless = List(`no-store`, `no-transform`, `only-if-cached`,
