+++ core/src/test/scala/org/http4s/headers/ContentTypeSpec.scala
@@ -1,13 +1,9 @@
 package org.http4s
-package parser
+package headers
 
-import org.http4s.headers.`Content-Type`
 import org.http4s.MediaType._
-import org.specs2.mutable.Specification
 
-class ContentTypeHeaderSpec extends Specification with HeaderParserHelper[`Content-Type`] {
-
-  def hparse(value: String): ParseResult[`Content-Type`] = HttpParser.CONTENT_TYPE(value)
+class ContentTypeSpec extends HeaderParserSpec(`Content-Type`) {
 
   def simple = `Content-Type`(`text/html`)
   def charset = `Content-Type`(`text/html`, Charset.`UTF-8`)
